feature: added mongodb standalone deployment
