+++ deploy/mongodb-standalone/mongodb-auth.yaml
@@ -14,30 +14,12 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-#
-[tool.poetry]
-name = "nuvolaris"
-version = "0.3.0"
-description = "Nuvolaris operator"
-authors = ["Michele Sciabarra <michele@sciabarra.com>"]
-
-[tool.poetry.dependencies]
-python = "^3.9"
-kopf = {extras = ["full-auth"], version = "^1.35.3"}
-PyYAML = "^6.0"
-pykube = "^0.15.0"
-Jinja2 = "^3.0.3"
-requests = "^2.27.1"
-flatdict = "^4.0.1"
-croniter = "^1.3.5"
-
-[tool.poetry.dev-dependencies]
-ipython = "^7.31.0"
-
-[tool.poetry.scripts]
-dbinit = "nuvolaris.couchdb:init"
-actionexecutor = "nuvolaris.actionexecutor:start"
-
-[build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
+---
+apiVersion: v1
+kind: Secret
+metadata:
+  name: mongodb-admin-auth
+  namespace: nuvolaris
+type: Opaque
+stringData:
+  mongodb_admin_psw: sUpErs3cr3t
