+++ tests/microk8s/whisk.yaml
@@ -77,6 +77,7 @@ spec:
     nuvolaris:
       user: nuvolaris
       password: s0meP@ass3
+    useOperator: False
   kafka:
     host: kafka
     volume-size: 10
